+++ src/models/extraParameters.ts
@@ -1,0 +1,3 @@
+export type ExtraParameters = {
+  [key: string]: string | number | null;
+};