+++ src/models/index.ts
@@ -1,2 +1,4 @@
 export * from "./filters/index";
 export * from "./operators";
+export * from "./sieveConstants";
+export * from "./extraParameters";