+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-06-12
+
+- Added possibility to add extra parameters:
+   `service.toValue(filter, {firstKey: "firstValue", secondKey: "secondValue"})`
+
 ## [1.0.1] - 2023-06-05
 
 - Changed some info for npmjs.org